# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: Help API
  version: "1.0"
servers:
  - url: http://192.168.25.107:8081
    description: Generated server url
paths:
...
  /categories/category/{categoryId}:
    get:
      tags:
        - category-controller
      operationId: getCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category'
    post:
      tags:
        - category-controller
      operationId: modifyUnit
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category'
...
  /tables/table/{tableId}:
    get:
      tags:
        - tables-controller
      operationId: getTable
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TablesDTO'
    post:
      tags:
        - tables-controller
      operationId: updateTable
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TablesDTO'
  /tables/{pageId}:
    get:
      tags:
        - tables-controller
      operationId: getAllTables
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: reserved
          in: query
          required: false
          schema:
            type: string
        - name: username
          in: query
          required: false
          schema:
            type: string
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageEatingPlace'
  /tables:
    post:
      tags:
        - tables-controller
      operationId: addNewTable
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TablesDTO'
  /tables/delete:
    post:
      tags:
        - tables-controller
      operationId: deleteTable
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: boolean
  /units/{pageId}:
    get:
      tags:
        - units-controller
      operationId: getUnits
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUnits'
  /units/unit/{unitId}:
    get:
      tags:
        - units-controller
      operationId: getUnit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Units'
    post:
      tags:
        - units-controller
      operationId: modifyUnit_1
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Units'
  /units:
    get:
      tags:
        - units-controller
      operationId: getUnitsList
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Units'
    post:
      tags:
        - units-controller
      operationId: addUnit
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Units'
...
  /users/user/{username}:
    get:
      tags:
        - user-controller
      operationId: getUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
components:
  schemas:
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int32
        category:
          type: string
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    Units:
      type: object
      properties:
        id:
          type: integer
          format: int32
        unitName:
          type: string
...
    TablesDTO:
      type: object
      properties:
        table_id:
          type: integer
          format: int32
        reserved:
          type: boolean
        waiter_name:
          type: string
        username:
          type: string
    EatingPlace:
      type: object
      properties:
        id:
          type: integer
          format: int32
        reserved:
          type: boolean
        waiterUsername:
          type: string
        waiterName:
          type: string
        active:
          type: boolean
    PageEatingPlace:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/EatingPlace'
        empty:
          type: boolean
    PageUnits:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Units'
        empty:
          type: boolean
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
        name:
          type: string
        lastname:
          type: string
        role:
          type: string
...